Approaching final, last version of adjacency with dictionaries
It was a fun way to do it, but storing face adjacency stats using dictionaries all the way through gets unnecessarily complex. It's harder to follow and work with than just using a numpy array as the output, rewriting shortly. That will make the error calculation much easier to improve, since it's currently quite basic.

Current version does not use EllipticEnvelope to estimate whether a dice fits within the simulated data.